+++ SourceryFramework/Sources/Parsing/SwiftSyntax/AST/Attribute+SwiftSyntax.swift
@@ -26,7 +26,7 @@ extension Attribute {
               case 1:
                   arguments["\(idx)"] = components[0].replacingOccurrences(of: "\"", with: "").trimmed as NSString
               default:
-                  assertionFailure("What is this")
+                  Log.astError("Unrecognized attribute format \(attribute.argument?.description ?? "")")
                   return
               }
           }